pc - fix bug in uploader script
A previous PR added the script, and there was a change
from the F20 version; we weren't clearing out the old
entries in the collection, so each time we ran the script,
we got duplicate entries (or triplicate if you ran it thrice, etc.)
We added a delete_many, but it was in the wrong place;
inside the loop, instead of at the top of the file.
So we ended up deleting each channels messages when
the next channels messages were loaded, resulting in
only the last channels messages sticking around.

This PR fixes that.